requirements.txt
diffusers==0.7.2
gradio==4.31.5
hydra-core==1.3.2
numpy==1.24.4
omegaconf==2.2.3
opencv_python==4.9.0.80
Pillow==9.5.0
Pillow==10.3.0
torch==1.13.1
tqdm==4.66.2
transformers==4.24.0
xformers==0.0.26.post1
